successfully rendering users listings. some styling also
