+++ client/src/assets/css/style.css
@@ -1607,6 +1607,7 @@ section {
     font-size: 20px;
     margin-left: 15%;
 
+
 }
 
 .center-button {
