+++ client/src/components/CategoryMenu/categorymenu.js
@@ -1,5 +1,5 @@
 import React from "react";
-//import { Link } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { useState, useEffect } from "react";
 import axios from 'axios';
