+++ client/src/components/MyAccount/myaccount.js
@@ -1,9 +1,34 @@
 import { Link } from 'react-router-dom';
-
+import { useState, useEffect } from "react";
+import Auth from '../../utils/auth';
 
 
 export default function MyAccount() {
+    // const [userId, setUserId] = useState();
 
+    // const getUserData = async () => {
+    //     try {
+    //         //use Auth obj to retrieve token from storage and set to variable 'token'
+    //         const token = Auth.loggedIn() ? Auth.getToken() : null;
+
+    //         if (!token) {
+    //             return false;
+    //         }
+        
+    //     //retrieve user id saved in token (user id)
+    //     const user = Auth.getProfile();
+    //         // console.log(user.data.id);
+    //         setUserId(user.data.id);
+            
+            
+    //         } catch (err) {
+    //             console.error(err);
+    //         }
+    //     };
+
+    //     useEffect(() => {
+    //         getUserData();
+    //     })
 
     return (
         <div>
@@ -33,7 +58,7 @@ export default function MyAccount() {
                                             <i className="bx bx-cube-alt"></i>
                                             <h4>My Listings</h4>
                                             <p>Click the button below to be redirected to your previous listings.</p>
-                                            <Link to='/mylistings' className="more-btn"> Here <i className="bx"></i></Link>
+                                            <Link to='/mylistings'  className="more-btn"> Here <i className="bx"></i></Link>
                                         </div>
                                     </div>
                                     <div className="col-xl-4 d-flex align-items-stretch" data-aos="zoom-in" data-aos-delay="300">
